Add support for a large-width image
Unfortunately, I do not know how to apply classes to an element in Markdown. Googling has led me to finding that the project would need a different parser to do so. However, pure HTML is allowed in Markdown, so this solution works if written like <img class="large-image" src="./img/azeroth_screenshot.png">